# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Oracle API
  description: AI-powered legal intelligence platform
  version: 1.0.0
  contact:
    name: Legal Oracle Team
    email: support@legaloracle.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://legal-oracle.netlify.app
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/v1/simulation/run:
    post:
      tags:
        - simulation
      summary: Simulate legal strategies
      description: Simulate legal strategies against AI opponents
      operationId: runSimulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRequest'
      responses:
        '200':
          description: Simulation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Gemini service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SimulationRequest:
      type: object
      required:
        - strategy
        - opponent_type
      properties:
        case_id:
          type: string
          nullable: true
        strategy:
          type: string
        opponent_type:
          type: string
          enum: [opposing_counsel, judge, ai_system]
        simulation_parameters:
          type: object
    SimulationResponse:
      type: object
      properties:
        simulation_id:
          type: string
        case_id:
          type: string
          nullable: true
        success_rate:
          type: number
        opponent_response:
          type: string
        key_insights:
          type: array
          items:
            type: string
        confidence_score:
          type: number
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
          description: Optional error code for programmatic handling